Refactor `atomWithFn` into util file
